feat(create-append-anything): encapsulate editor actions
