+++ lib/features/create-append-anything/index.js
@@ -5,6 +5,7 @@ import ContextPadModule from 'diagram-js/lib/features/context-pad';
 
 import CreateMenuProvider from './CreateMenuProvider';
 import CreatePaletteProvider from './CreatePaletteProvider';
+import CreateAppendEditorActions from './CreateAppendEditorActions';
 import AppendMenuProvider from './AppendMenuProvider';
 import AppendContextPadProvider from './AppendContextPadProvider';
 
@@ -18,11 +19,13 @@ export default {
   __init__: [
     'createMenuProvider',
     'createPaletteProvider',
+    'createAppendEditorActions',
     'appendMenuProvider',
     'appendContextPadProvider'
   ],
   createMenuProvider: [ 'type', CreateMenuProvider ],
   createPaletteProvider: [ 'type', CreatePaletteProvider ],
+  createAppendEditorActions: [ 'type', CreateAppendEditorActions ],
   appendMenuProvider: [ 'type', AppendMenuProvider ],
   appendContextPadProvider: [ 'type', AppendContextPadProvider ]
 };
