+++ lib/features/editor-actions/BpmnEditorActions.js
@@ -10,7 +10,6 @@ import {
   getBBox
 } from 'diagram-js/lib/util/Elements';
 
-
 /**
  * Registers and executes BPMN specific editor actions.
  *
@@ -178,12 +177,6 @@ BpmnEditorActions.prototype._registerDefaultActions = function(injector) {
   if (selection && contextPad) {
     this._registerAction('replaceElement', function(event) {
       contextPad.triggerEntry('replace', 'click', event);
-    });
-  }
-
-  if (selection && contextPad) {
-    this._registerAction('appendElement', function(event) {
-      contextPad.triggerEntry('append', 'click', event);
     });
   }
