+++ test/spec/features/editor-actions/BpmnEditorActionsSpec.js
@@ -281,73 +281,4 @@ describe('features/editor-actions', function() {
 
   });
 
-
-  describe('#appendElement', function() {
-
-    beforeEach(bootstrapModeler(basicXML, {
-      modules: [
-        selectionModule,
-        bpmnEditorActionsModule,
-        modelingModule,
-        coreModule,
-        contextPad
-      ]
-    }));
-
-
-    it('should open append element', inject(function(elementRegistry, selection, editorActions, eventBus) {
-
-      // given
-      const element = elementRegistry.get('StartEvent_1');
-
-      selection.select(element);
-      var changedSpy = sinon.spy();
-
-      // when
-      eventBus.once('popupMenu.open', changedSpy);
-
-      editorActions.trigger('appendElement', {});
-
-      // then
-      expect(changedSpy).to.have.been.called;
-    }));
-
-
-    it('should not open append element if no selection', inject(function(editorActions, eventBus) {
-
-      // given
-      var changedSpy = sinon.spy();
-
-      // when
-      eventBus.once('popupMenu.open', changedSpy);
-
-      editorActions.trigger('appendElement', {});
-
-      // then
-      expect(changedSpy).to.not.have.been.called;
-    }));
-
-
-    it('should not open append element if multiple elements selected', inject(function(elementRegistry, selection, editorActions, eventBus) {
-
-      // given
-      var elementIds = [ 'StartEvent_1', 'UserTask_1' ];
-      var elements = elementIds.map(function(id) {
-        return elementRegistry.get(id);
-      });
-
-      selection.select(elements);
-      var changedSpy = sinon.spy();
-
-      // when
-      eventBus.once('popupMenu.open', changedSpy);
-
-      editorActions.trigger('appendElement', {});
-
-      // then
-      expect(changedSpy).to.not.have.been.called;
-    }));
-
-  });
-
 });
